+++ CollapsibleUI.plugin.js
@@ -3,7 +3,7 @@
  * @author TenorTheHusky
  * @authorId 563652755814875146
  * @description A feature-rich BetterDiscord plugin that reworks the Discord UI to be significantly more modular
- * @version 8.4.9
+ * @version 8.5.0
  * @donate https://ko-fi.com/benjaminpryor
  * @patreon https://www.patreon.com/BenjaminPryor
  * @website https://github.com/programmer2514/BetterDiscord-CollapsibleUI
@@ -21,18 +21,19 @@ module.exports = (() => {
         github_username: 'programmer2514',
       },
       ],
-      version: '8.4.9',
+      version: '8.5.0',
       description: 'A feature-rich BetterDiscord plugin that reworks the Discord UI to be significantly more modular',
       github: 'https://github.com/programmer2514/BetterDiscord-CollapsibleUI',
       github_raw: 'https://raw.githubusercontent.com/programmer2514/BetterDiscord-CollapsibleUI/main/CollapsibleUI.plugin.js',
     },
     changelog: [{
-      title: '8.4.9',
+      title: '8.5.0',
       items: [
-        'Fixed Members List occasionally reversing itself',
+        'Fixed profile panel and settings buttons on latest update',
+        'Fixed plugin failing to load in thread channels',
       ],
     }, {
-      title: '1.0.0 - 8.4.8',
+      title: '1.0.0 - 8.4.9',
       items: [
         `See the full changelog here:\
            https://programmer2514.github.io/?l=cui-changelog`,
@@ -458,8 +459,8 @@ module.exports = (() => {
         this.classMembersListWrapper = 'container_cbd271';
         this.classMembersListMember = 'member_a31c43';
         this.classMembersListInner = 'members_cbd271';
-        this.classProfilePanel = 'userPanelOuter_c69a7b';
-        this.classProfilePanelWrapper = 'profilePanel_b433b4';
+        this.classProfilePanel = 'inner_c69a7b';
+        this.classProfilePanelWrapper = 'outer_c69a7b';
         this.classTextInput = '[data-slate-string="true"]';
         this.classNoChat = 'noChat_a7d72e';
         this.classMsgButtons = 'wrapper_f7e168';
@@ -469,6 +470,7 @@ module.exports = (() => {
         this.classUnreadDmBadgeEyebrow = 'eyebrow_df8943';
         this.classUnreadDmBadgeShape = 'baseShapeRound_df8943';
         this.classUnreadDmBadgeLocation = 'unreadMentionsIndicatorTop_fea3ef';
+        this.classContentWindow = 'container_a6d69a';
       }
 
       if (BdApi.Plugins.isEnabled('ChannelDms')
@@ -489,8 +491,7 @@ module.exports = (() => {
       this.userArea = document.querySelector('.panels_a4d4d9');
       this.profilePanel = document.querySelector('.'
         + this.classProfilePanel);
-      this.profilePanelInner = document.querySelector('.userPanelInner_c69a7b')
-        ?.firstElementChild;
+      this.profilePanelInner = this.profilePanel?.firstElementChild;
       this.profilePanelFooter = document.querySelector('.footer_a823cd');
       this.profilePanelWrapper = document.querySelector('.'
         + this.classProfilePanelWrapper);
@@ -529,12 +530,12 @@ module.exports = (() => {
         + ' 2v3a1 1 0 1 0 2 0V6a4 4 0 0 0-4-4h-3a1 1 0 1 0 0 2h3ZM20 18a2 2 0 0'
         + ' 1-2 2h-3a1 1 0 1 0 0 2h3a4 4 0 0 0 4-4v-3a1 1 0 1 0-2 0v3Z"]')
         ?.parentElement.parentElement.parentElement;
-      this.msgBarBtnContainer = document.querySelector('.buttons_d0696b');
+      this.msgBarBtnContainer = document.querySelector('.buttons_bdf0de');
       this.membersListInner = document.querySelector('.' + this.classMembersListInner);
       this.membersListWrapper = document.querySelector('.container_cbd271');
       this.contentWindow = document.querySelector('.chatContent_a7d72e');
       if (!this.contentWindow)
-        this.contentWindow = document.querySelector('.container_a6d69a');
+        this.contentWindow = document.querySelector('.' + this.classContentWindow);
 
       this.callContainerExists = (document.querySelector('.'
         + this.classCallContainer));
@@ -2637,6 +2638,7 @@ module.exports = (() => {
             this.profilePanel.style.maxHeight = '100%';
             this.profilePanel.style.height = '100%';
             this.profilePanel.style.right = '0';
+            this.profilePanel.style.background = 'var(--background-secondary-alt)';
           }
           else {
             this.profilePanel.style.removeProperty('position');
@@ -2644,6 +2646,7 @@ module.exports = (() => {
             this.profilePanel.style.removeProperty('max-height');
             this.profilePanel.style.removeProperty('height');
             this.profilePanel.style.removeProperty('right');
+            this.profilePanel.style.removeProperty('background');
           }
           break;
 
@@ -3097,6 +3100,7 @@ module.exports = (() => {
               || mutationList[i].addedNodes[0]?.classList?.contains(cui.classProfilePanelWrapper)
               || mutationList[i].addedNodes[0]?.classList?.contains(cui.classCallContainer)
               || mutationList[i].addedNodes[0]?.classList?.contains(cui.classMembersListInner)
+              || mutationList[i].addedNodes[0]?.classList?.contains(cui.classContentWindow)
               || mutationList[i].removedNodes[0]?.classList?.contains(cui.classCallContainer)) {
               cui.initialize();
               return;